Add files via upload
Here is the final list for the remaining $TOPIA token being distributed from unclaimed worlds as we mentioned prior. This will close out the remainder of the World Owners Claim from tokenomics.

We took the snapshot for this at block 18810043

Attached are the distributions. These $TOPIA tokens will be distributed on the HYTOPIA Chain.

These will be going out within the next few days.